+++ testing/node-driver/src/main/kotlin/net/corda/testing/node/MockServices.kt
@@ -1,6 +1,7 @@
 package net.corda.testing.node
 
 import com.google.common.collect.MutableClassToInstanceMap
+import net.corda.core.contracts.Attachment
 import net.corda.core.contracts.ContractClassName
 import net.corda.core.contracts.StateRef
 import net.corda.core.cordapp.CordappProvider
@@ -39,12 +40,16 @@ import net.corda.testing.internal.MockCordappProvider
 import net.corda.testing.internal.configureDatabase
 import net.corda.testing.node.internal.*
 import net.corda.testing.services.MockAttachmentStorage
+import java.io.ByteArrayOutputStream
 import java.security.KeyPair
 import java.sql.Connection
 import java.time.Clock
 import java.time.Instant
 import java.util.*
 import java.util.function.Consumer
+import java.util.jar.JarFile
+import java.util.zip.ZipEntry
+import java.util.zip.ZipOutputStream
 import javax.persistence.EntityManager
 
 /** Returns a simple [IdentityService] containing the supplied [identities]. */
@@ -142,6 +147,24 @@ open class MockServices private constructor(
 
         // Because Kotlin is dumb and makes not publicly visible objects public, thus changing the public API.
         private val mockStateMachineRecordedTransactionMappingStorage = MockStateMachineRecordedTransactionMappingStorage()
+
+        private val dummyAttachment by lazy {
+            val inputStream = ByteArrayOutputStream().apply {
+                ZipOutputStream(this).use {
+                    with(it) {
+                        putNextEntry(ZipEntry(JarFile.MANIFEST_NAME))
+                    }
+                }
+            }.toByteArray().inputStream()
+            val attachment = object : Attachment {
+                override val id get() = throw UnsupportedOperationException()
+                override fun open() = inputStream
+                override val signerKeys get() = throw UnsupportedOperationException()
+                override val signers: List<Party> get() = throw UnsupportedOperationException()
+                override val size: Int = 512
+            }
+            attachment
+        }
     }
 
     private class MockStateMachineRecordedTransactionMappingStorage : StateMachineRecordedTransactionMappingStorage {
@@ -216,6 +239,11 @@ open class MockServices private constructor(
     @JvmOverloads
     constructor(cordappPackages: List<String>, initialIdentityName: CordaX500Name, identityService: IdentityService, networkParameters: NetworkParameters)
             : this(cordappPackages, TestIdentity(initialIdentityName), identityService, networkParameters)
+
+
+    @JvmOverloads
+    constructor(cordappPackages: List<String>, initialIdentityName: CordaX500Name, identityService: IdentityService, networkParameters: NetworkParameters, key: KeyPair)
+            : this(cordappPackages, TestIdentity(initialIdentityName, key), identityService, networkParameters)
 
     /**
      * A helper constructor that requires at least one test identity to be registered, and which takes the package of
@@ -321,6 +349,8 @@ open class MockServices private constructor(
 
     override fun loadState(stateRef: StateRef) = servicesForResolution.loadState(stateRef)
     override fun loadStates(stateRefs: Set<StateRef>) = servicesForResolution.loadStates(stateRefs)
+
+    override fun loadContractAttachment(stateRef: StateRef, forContractClassName: ContractClassName?) = try { servicesForResolution.loadContractAttachment(stateRef) } catch (e: Exception) { dummyAttachment }
 }
 
 /**