+++ testing/node-driver/src/main/kotlin/net/corda/testing/node/TestCordapp.kt
@@ -1,6 +1,7 @@
 package net.corda.testing.node
 
 import net.corda.core.DoNotImplement
+import net.corda.core.cordapp.DEFAULT_CORDAPP_VERSION
 import net.corda.core.internal.PLATFORM_VERSION
 import net.corda.testing.node.internal.TestCordappImpl
 import net.corda.testing.node.internal.simplifyScanPackages
@@ -26,6 +27,9 @@ interface TestCordapp {
     /** Returns the target platform version, defaults to the current platform version if not specified. */
     val targetVersion: Int
 
+    /** Returns the cordapp version. */
+    val cordappVersion: String
+
     /** Returns the config for this CorDapp, defaults to empty if not specified. */
     val config: Map<String, Any>
 
@@ -34,9 +38,6 @@ interface TestCordapp {
 
     /** Returns whether the CorDapp should be jar signed. */
     val signJar: Boolean
-
-    /** Returns the contract version, default to 1 if not specified. */
-    val cordaContractVersion: Int
 
     /** Return a copy of this [TestCordapp] but with the specified name. */
     fun withName(name: String): TestCordapp
@@ -60,7 +61,7 @@ interface TestCordapp {
      *  Optionally can pass in the location of an existing java key store to use */
     fun signJar(keyStorePath: Path? = null): TestCordappImpl
 
-    fun withCordaContractVersion(version: Int): TestCordappImpl
+    fun withCordappVersion(version: String): TestCordappImpl
 
     class Factory {
         companion object {
@@ -83,6 +84,7 @@ interface TestCordapp {
                         vendor = "test-vendor",
                         title = "test-title",
                         targetVersion = PLATFORM_VERSION,
+                        cordappVersion = DEFAULT_CORDAPP_VERSION.toString(),
                         config = emptyMap(),
                         packages = simplifyScanPackages(packageNames),
                         classes = emptySet()