+++ docs/source/changelog.rst
@@ -7,6 +7,10 @@ release, see :doc:`upgrade-notes`.
 Unreleased
 ----------
 
+* Transaction building and verification enforces new contract attachment version non-downgrade rule.
+  For a given contract class, the contract attachment of the output states must be of the same or newer version than the contract attachment of the input states.
+  See :ref:`Contract attachment non-downgrade rule <contract_non-downgrade_rule_ref>` for further information.
+
 * Automatic Constraints propagation for hash-constrained states to signature-constrained states.
   This allows Corda 4 signed CorDapps using signature constraints to consume existing hash constrained states generated
   by unsigned CorDapps in previous versions of Corda.